공지사항 관련 page 수정 #75
