+++ src/main/java/com/tsp/new_tsp_front/api/notice/service/impl/FrontNoticeJpaService.java
@@ -1,32 +1,29 @@
 package com.tsp.new_tsp_front.api.notice.service.impl;
 
 import com.tsp.new_tsp_front.api.notice.domain.FrontNoticeDTO;
+import com.tsp.new_tsp_front.api.notice.domain.FrontNoticeEntity;
+import com.tsp.new_tsp_front.exception.TspException;
 import lombok.RequiredArgsConstructor;
-import org.springframework.cache.annotation.Cacheable;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.Map;
 
+import static com.tsp.new_tsp_front.exception.ApiExceptionType.NOT_FOUND_NOTICE;
+
 
 @Service
 @RequiredArgsConstructor
 public class FrontNoticeJpaService {
+    private final FrontNoticeJpaQueryRepository frontNoticeJpaQueryRepository;
     private final FrontNoticeJpaRepository frontNoticeJpaRepository;
 
-    /**
-     * <pre>
-     * 1. MethodName : findNoticeCount
-     * 2. ClassName  : FrontNoticeJpaApiService.java
-     * 3. Comment    : 프론트 > 공지사항 리스트 갯수 조회
-     * 4. 작성자       : CHO
-     * 5. 작성일       : 2022. 08. 16.
-     * </pre>
-     */
-    @Transactional(readOnly = true)
-    public int findNoticeCount(Map<String, Object> noticeMap) {
-        return frontNoticeJpaRepository.findNoticeCount(noticeMap);
+    private FrontNoticeEntity oneNotice(Long idx) {
+        return frontNoticeJpaRepository.findById(idx)
+                .orElseThrow(() -> new TspException(NOT_FOUND_NOTICE));
     }
 
     /**
@@ -38,39 +35,9 @@ public class FrontNoticeJpaService {
      * 5. 작성일      : 2022. 08. 16.
      * </pre>
      */
-    @Cacheable(value = "notice", key = "#noticeMap")
     @Transactional(readOnly = true)
-    public List<FrontNoticeDTO> findNoticeList(Map<String, Object> noticeMap) {
-        return frontNoticeJpaRepository.findNoticeList(noticeMap);
-    }
-
-    /**
-     * <pre>
-     * 1. MethodName : findFixedNoticeCount
-     * 2. ClassName  : FrontNoticeJpaApiService.java
-     * 3. Comment    : 프론트 > 상단 고정 공지사항 리스트 갯수 조회
-     * 4. 작성자      : CHO
-     * 5. 작성일      : 2022. 09. 24.
-     * </pre>
-     */
-    @Transactional(readOnly = true)
-    public int findFixedNoticeCount(Map<String, Object> noticeMap) {
-        return frontNoticeJpaRepository.findNoticeCount(noticeMap);
-    }
-
-    /**
-     * <pre>
-     * 1. MethodName : findFixedNoticeList
-     * 2. ClassName  : FrontNoticeJpaService.java
-     * 3. Comment    : 프론트 > 상단 고정 공지사항 리스트 조회
-     * 4. 작성자      : CHO
-     * 5. 작성일      : 2022. 09. 24.
-     * </pre>
-     */
-    @Cacheable(value = "notice", key = "#noticeMap.get('topFixed')")
-    @Transactional(readOnly = true)
-    public List<FrontNoticeDTO> findFixedNoticeList(Map<String, Object> noticeMap) {
-        return frontNoticeJpaRepository.findNoticeList(noticeMap);
+    public Page<FrontNoticeDTO> findNoticeList(Map<String, Object> noticeMap, PageRequest pageRequest) {
+        return frontNoticeJpaQueryRepository.findNoticeList(noticeMap, pageRequest);
     }
 
     /**
@@ -82,10 +49,9 @@ public class FrontNoticeJpaService {
      * 5. 작성일      : 2022. 08. 16.
      * </pre>
      */
-    @Cacheable(value = "notice", key = "#idx")
     @Transactional(readOnly = true)
     public FrontNoticeDTO findOneNotice(Long idx) {
-        return this.frontNoticeJpaRepository.findOneNotice(idx);
+        return FrontNoticeEntity.toDto(oneNotice(idx));
     }
 
     /**
@@ -99,7 +65,7 @@ public class FrontNoticeJpaService {
      */
     @Transactional(readOnly = true)
     public FrontNoticeDTO findPrevOneNotice(Long idx) {
-        return this.frontNoticeJpaRepository.findPrevOneNotice(idx);
+        return this.frontNoticeJpaQueryRepository.findPrevOneNotice(idx);
     }
 
     /**
@@ -113,6 +79,6 @@ public class FrontNoticeJpaService {
      */
     @Transactional(readOnly = true)
     public FrontNoticeDTO findNextOneNotice(Long idx) {
-        return this.frontNoticeJpaRepository.findNextOneNotice(idx);
+        return this.frontNoticeJpaQueryRepository.findNextOneNotice(idx);
     }
 }
